+++ lib/data/models/database.dart
@@ -1,7 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
-import 'package:notion_client/data/models/base_object.dart';
-
 import 'emoji_icon.dart';
+import 'interfaces/base_object.dart';
 part 'database.g.dart';
 
 @JsonSerializable()
