+++ lib/data/models/interfaces/base_object.dart
